hardware/ril: Fix rild crash when phone app gets killed continuously
Whenever phone app gets killed, the listen events are getting piled up in
watch table, thus not allowing further fd command events to be registered
which is resulting in rild crash. As part of DSDS changes, listen event was
made persistent. So, it needs to be deleted and then added to the watch list
whenever connection is reset.

Change-Id: I9d454e2269eaa3b5db6951a6b58dab688c23933c
CRs-Fixed: 283561